Use patching technique for OmniboxViewViews
