+++ chromium_src/chrome/browser/ui/views/location_bar/location_bar_view.cc
@@ -4,15 +4,12 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "brave/browser/ui/omnibox/brave_omnibox_client_impl.h"
-#include "brave/browser/ui/views/omnibox/brave_omnibox_view_views.h"
 
 #define BRAVE_LAYOUT_TRAILING_DECORATIONS     \
   if (right_most && right_most->GetVisible()) \
     trailing_decorations.AddDecoration(0, height(), false, 0, 0, right_most);
 
 #define ChromeOmniboxClient BraveOmniboxClientImpl
-#define OmniboxViewViews BraveOmniboxViewViews
 #include "../../../../../../../chrome/browser/ui/views/location_bar/location_bar_view.cc"  // NOLINT
 #undef ChromeOmniboxClient
-#undef OmniboxViewViews
 #undef BRAVE_LAYOUT_TRAILING_DECORATIONS
